split body into a seperate file added nessisary getitem method to system made str, and repr more reasonable
